refactor: Improve settings/state storing
Previously there was a flash of the default state on the initial load. Now, the default state is set to a promise which resolves to the state on disk.

The stores/ directory has been removed and everything is in Root.tsx now due to order of execution problems. (TAURI_STORE) wasn't initialized yet